Changed to use react 18 root api to enable concurrent features
