+++ src/components/index.js
@@ -1,10 +1,8 @@
-import Landing from './Landing';
 import NavBar from './NavBar';
 import Footer from './Footer';
 import ItemList from './ItemList';
 
 export {
-  Landing,
   NavBar,
   Footer,
   ItemList,
